Create IGalleryEntry.cs
* Moved the new IGalleryEntry interface from the GalleryEntry.cs file to its own file.
