+++ src/VS2022/ExtensionManager2022.csproj
@@ -49,6 +49,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Models\GalleryEntry.cs" />
+    <Compile Include="Models\IGalleryEntry.cs" />
     <Compile Include="source.extension.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
